refactor : 상품 목록 최종 페이지 판별 로직 수정
